# public/swagger.json
{
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Документация Golibrary API.",
    "title": "Golibrary API.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/library/authors": {
      "get": {
        "description": "Получение списка авторов",
        "tags": [
          "author"
        ],
        "operationId": "Author",
        "responses": {
          "200": {
            "description": "Cписок авторов успешно получен",
            "schema": {
              "$ref": "#/definitions/Author"
            }
          }
        }
      }
    },
    "/library/return/{bookId}/{userId}": {
      "put": {
        "description": "Сдать книгу",
        "tags": [
          "book"
        ],
        "operationId": "BookReturn",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "BookId",
            "name": "bookId",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "UserId",
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Книга успешно сдана",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        }
      }
    },
    "/library/take/{bookId}/{userId}": {
      "put": {
        "description": "Взять книгу",
        "tags": [
          "book"
        ],
        "operationId": "BookTake",
        "parameters": [
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "BookId",
            "name": "bookId",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "format": "int64",
            "x-go-name": "UserId",
            "name": "userId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Книга успешно получена",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        }
      }
    },
    "/library/users": {
      "get": {
        "description": "Список читателей",
        "tags": [
          "user"
        ],
        "operationId": "UsersList",
        "responses": {
          "200": {
            "description": "Список читателей успешно получен",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Author": {
      "type": "object",
      "properties": {
        "birth_date": {
          "type": "string",
          "x-go-name": "BirthDate"
        },
        "books": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Book"
          },
          "x-go-name": "Books"
        },
        "id": {
          "description": "in:body",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "golibrary/docs"
    },
    "Book": {
      "type": "object",
      "properties": {
        "author_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "AuthorID"
        },
        "available": {
          "type": "boolean",
          "x-go-name": "Available"
        },
        "id": {
          "description": "in:body",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "title": {
          "type": "string",
          "x-go-name": "Title"
        },
        "user_id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UserID"
        }
      },
      "x-go-name": "BookReturn",
      "x-go-package": "golibrary/internal/model"
    },
    "User": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "id": {
          "description": "in:body",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-name": "UsersList",
      "x-go-package": "golibrary/docs"
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}